Rename mailMerge.js -> group-merge.js
