+++ src/group-merge.js
